requirements.txt
torch
torchaudio
torchvision
gradio
diffusers
transformers
openai-whisper
ffmpeg-python
